+++ backend/middleware/file.js
@@ -1,0 +1,25 @@
+const multer = require("multer");
+
+const MIME_TYPE_MAP = {
+    'image/png': 'png',
+    'image/jpeg': 'jpg',
+    'image/jpg': 'jpg'
+}
+
+const storage = multer.diskStorage({
+    destination: (req, file, cb) => {
+        const isValid = MIME_TYPE_MAP[file.mimetype];
+        error = new Error(" Invalid mime type");
+        if (isValid){
+            error = null;
+        }
+        cb(error, "backend/images");
+    },
+    filename: (req, file, cb) => {
+        const nombre = file.originalname.toLowerCase().split(' ').join('-');
+        const ext = MIME_TYPE_MAP[file.mimetype];
+        cb(null, nombre + '-' + Date.now() + '.' + ext);
+    }
+});
+
+module.exports = multer({storage: storage}).single("image");
