+++ backend/routes/user.js
@@ -1,53 +1,9 @@
 const express = require("express");
-//const bcrypt = require("bcrypt");
-const User = require("../models/user");
-const jwt = require("jsonwebtoken");
 
+const UserController = require("../controllers/user");
 const router = express.Router();
 
-router.post("/signup", (req, res, next) => {
-    console.log(" entrando en signup en el servidor ");
-    const user = new User({
-        email: req.body.email,
-        password: req.body.password
-    });
-    user.save().then(result => {
-        res.status(201).json({
-            message: 'User created!',
-            result: result
-        })
-    });
-});
-
-
-router.post("/login", (req, res, next) => {
-    let fetchedUser;
-    User.findOne({ email: req.body.email})
-    .then(user => {
-        if (!user){
-            return res.status(401).json({
-                message: "Auth failed!"
-            });
-        }
-        fetchedUser = user;
-        return req.body.password == user.password;
-    }).then(result => {
-        const token = jwt.sign({email: fetchedUser.email, userId: fetchedUser._id}, 
-            'secreto_creador_token',
-            {expiresIn: "1h"});
-        
-        res.status(200).json({
-            token: token,
-            expiresIn: "3600",
-            userId: fetchedUser._id
-        });
-
-    }).catch(error => {
-        return res.status(401).json({
-            message: "Auth failed!"
-        });
-    });
-});
-
+router.post("/signup", UserController.createUser);
+router.post("/login", UserController.userLogin);
 
 module.exports = router;
