+++ src/app/angular-material.module.ts
@@ -1,0 +1,34 @@
+import { NgModule } from '@angular/core';
+
+import { MatInputModule, 
+        MatCardModule, 
+        MatButtonModule, 
+        MatToolbarModule, 
+        MatExpansionModule, 
+        MatProgressSpinnerModule, 
+        MatPaginatorModule, 
+        MatDialogModule} from '@angular/material';
+
+@NgModule({
+    imports: [
+        MatInputModule,
+        MatCardModule,
+        MatButtonModule,
+        MatToolbarModule,
+        MatExpansionModule,
+        MatProgressSpinnerModule,
+        MatPaginatorModule,
+        MatDialogModule
+    ],
+    exports: [
+        MatInputModule,
+        MatCardModule,
+        MatButtonModule,
+        MatToolbarModule,
+        MatExpansionModule,
+        MatProgressSpinnerModule,
+        MatPaginatorModule,
+        MatDialogModule
+    ]
+})
+export class AngularMaterialModule {}
