+++ src/app/posts/posts.module.ts
@@ -1,0 +1,21 @@
+import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
+import { AngularMaterialModule } from '../angular-material.module';
+import { PostCreateComponent } from './post-create/post-create.component';
+import { PostListComponent } from './post-list/post-list.component';
+
+@NgModule({
+    declarations: [
+        PostCreateComponent,
+        PostListComponent
+    ],
+    imports: [
+        CommonModule,
+        ReactiveFormsModule,
+        RouterModule,
+        AngularMaterialModule
+    ]
+})
+export class PostsModule {}
